Move GetAlternativeBeaconKeys() function to beacon.cpp
