+++ src/main.cpp
@@ -4557,29 +4557,6 @@ std::string RetrieveMd5(std::string s1)
     }
 }
 
-std::set<std::string> GetAlternativeBeaconKeys(const std::string& cpid)
-{
-    int64_t iMaxSeconds = 60 * 24 * 30 * 6 * 60;
-    std::set<std::string> result;
-
-    for(const auto& item : ReadCacheSection(Section::BEACONALT))
-    {
-        const std::string& key = item.first;
-        const std::string& value = item.second.value;
-        if(!std::equal(cpid.begin(), cpid.end(), key.begin()))
-            continue;
-
-        const int64_t iAge = pindexBest != NULL
-            ? pindexBest->nTime - item.second.timestamp
-            : 0;
-        if (iAge > iMaxSeconds)
-            continue;
-
-        result.emplace(value);
-    }
-    return result;
-}
-
 bool IsCPIDValidv2(MiningCPID& mc, int height)
 {
     //09-25-2016: Transition to CPID Keypairs.
